requirements.txt
gradio>=4.0.0
langchain>=0.2.0
langchain-anthropic>=0.1.15
langgraph>=0.1.0
python-dotenv>=1.0.1
langchain-openai>=0.1.0
langchain-google-genai>=1.0.0
